Move `Emphasis` and `EmphasisStyle` to Files
